Fixes a set of overflow bugs in hmmpgmd that Simon reported. A 32-bit integer (seq_count) was being multiplied by 10 to compute the size of a datastructure field. While the variable containing the size was 64-bits, seq_count * 10 was being computed using 32-bit arithmetic.  With the latest EBI database file, this computation overflowed and created a negative value, which was then sign-extended to 64 bits and assigned to the size variable.
All variables involved in this computation are now 64 bits.

+++ src/cachedb.c
@@ -39,7 +39,7 @@ p7_seqcache_Open(char *seqfile, P7_SEQCACHE **ret_cache, char *errbuf)
   int                val;
   int                status;
 
-  int32_t            seq_cnt;
+  uint64_t            seq_cnt;
   int32_t            db_cnt;
   int32_t            db_inx[32];
   uint32_t           db_key;
